feature: insert image into editor from editor menu.
- Media file upload fn

+++ src/stores/file.ts
@@ -1,3 +1,4 @@
+import { useAuthStore } from './auth';
 import { defineStore } from 'pinia';
 import { sdk } from 'src/boot/axios';
 import { repositories } from 'src/boot/repositories';
@@ -18,18 +19,58 @@ export const useFileStore = defineStore('fileStore', () => {
     repositories.files.deleteByName(name);
   };
 
+  const authStore = useAuthStore();
+
   const preserveFiles = async () => {
+    if (authStore.user?.isAnonymous) {
+      return;
+    }
+
     let firstFile = await repositories.files.getFirst();
     while (firstFile) {
       if (!firstFile) {
         return;
       }
-      const rspns = sdk.files.uploadFile(firstFile);
-      console.log('✎: [line 24][file.ts] rspns: ', rspns);
-      await deleteFile(firstFile.name);
-      await sleep(1500);
-      firstFile = await repositories.files.getFirst();
+      try {
+        await sdk.files.uploadFile(firstFile);
+        await deleteFile(firstFile.name);
+        await sleep(1500);
+        firstFile = await repositories.files.getFirst();
+      } catch (e) {
+        console.error(e);
+        preserveFiles();
+      }
     }
+  };
+
+  const uploadMediaFile = async (): Promise<string> => {
+    // Programmatically create input for file upload (image extensions)
+    const input = document.createElement('input');
+    input.type = 'file';
+    input.accept = 'image/*';
+    input.multiple = false;
+
+    const abortController = new AbortController();
+
+    const fileUploadPromise = new Promise<string>((resolve) => {
+      input.addEventListener(
+        'change',
+        async () => {
+          if (!input.files?.length) {
+            return;
+          }
+          const file = input.files[0];
+          await saveFile(file);
+          resolve(file.name);
+          abortController.abort();
+          console.log('✎: [line 70][import-store.ts] file: ', file);
+        },
+        { signal: abortController.signal }
+      );
+    });
+
+    input.click();
+    return fileUploadPromise;
   };
 
   return {
@@ -37,5 +78,6 @@ export const useFileStore = defineStore('fileStore', () => {
     getFile,
     deleteFile,
     preserveFiles,
+    uploadMediaFile,
   };
 });
